+++ Emuera/UI/Game/StringStyle.cs
@@ -1,4 +1,4 @@
-﻿using MinorShift.Emuera.Runtime.Config;
+using MinorShift.Emuera.Runtime.Config;
 using System.Drawing;
 
 namespace MinorShift.Emuera.UI.Game;
@@ -42,7 +42,19 @@ internal struct StringStyle
 	public Color? BackgroundColor;
 	public Color ButtonColor;
 	public bool ColorChanged;
-	public FontStyle FontStyle;
+	FontStyle _fontStyle;
+	public FontStyle FontStyle
+	{
+		readonly get { return _fontStyle; }
+		set
+		{
+			_fontStyle = value;
+			HasUnderline = _fontStyle.HasFlag(FontStyle.Underline);
+			HasStrikeout = _fontStyle.HasFlag(FontStyle.Strikeout);
+		}
+	}
+	public bool HasStrikeout;
+	public bool HasUnderline;
 	public string Fontname;
 	public override bool Equals(object? obj)
 	{